+++ CHANGES.md
@@ -9,6 +9,7 @@
 * move capabilities check into `class Init`
 * move API classes to subdirectory
 * move WP-CLI classes to subdirectory
+* refactor autoloader to grab all subdirectories
 
 #### 7.4.4 / 2017-11-29
 * fixed bug in remote install where Bitbucket credentials weren't transferred to Basic_Auth_Loader, [#630](https://github.com/afragen/github-updater/issues/630)